requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
pandas>=1.3.0
numpy>=1.20.0
Pillow>=8.0.0
tqdm>=4.62.0

